software/axiuart_driver/requirements.txt
# AXIUART Driver Requirements
# Core communication dependencies
pyserial>=3.5
typing-extensions>=4.0

# GUI dependencies (tkinter usually included with Python)
# Additional GUI libraries for advanced features
matplotlib>=3.5.0  # For plotting and data visualization
numpy>=1.21.0      # For numerical operations

# Development and testing dependencies  
pytest>=7.0.0